printf support 64bits value.

+++ include/printf.h
@@ -6,9 +6,9 @@
 #define va_arg(v,l)	__builtin_va_arg(v,l)
 typedef __builtin_va_list va_list;
 
-u32 vsnprintf(char *buf, u32 size, const char *format, va_list vl);
-u32 snprintf(char *buf, u32 size, const char *format, ...);
-u32 printf(const char *format, ...);
+usize vsnprintf(char *buf, u32 size, const char *format, va_list vl);
+usize snprintf(char *buf, u32 size, const char *format, ...);
+usize printf(const char *format, ...);
 void print_bits(usize size, void *ptr);
 int atoi(const char *str);
